add display code
This uses a 128x64 SPI OLED with SH1106 controller.
Note that for unknown reasons, I was not able to get an i2c display
working as a second i2c bus (in addition to es8388), as the i2c buses
seemed to interfere with each other (only the last initialized would
work). This i s why i switched to a SPI display ;-)